% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_readme.R
\name{add_readme}
\alias{add_readme}
\title{Add README to Project}
\usage{
add_readme(project_path, project_title, project_description)
}
\arguments{
\item{project_path}{The path to the project directory.}

\item{project_title}{The title of the project.}

\item{project_description}{A brief description of the project.}
}
\value{
None. The function writes a README.md file to the project directory.
}
\description{
Copies a README template to the specified project directory and customizes it.
}
