% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_example_files.R
\name{create_example_files}
\alias{create_example_files}
\title{Create Example Files}
\usage{
create_example_files(project_path)
}
\arguments{
\item{project_path}{The path to the project directory.}
}
\value{
No return value, writes files to disk.
}
\description{
Generates template files within the newly created R project.
}
\details{
This function populates example files such as \code{install.R}, \code{main.R}, example functions, scripts, and a Quarto report.
}
