% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_project_structure.R
\name{create_project_structure}
\alias{create_project_structure}
\title{Create Project Structure}
\usage{
create_project_structure(project_path)
}
\arguments{
\item{project_path}{The path where the project structure should be created.}
}
\value{
No return value, creates directories.
}
\description{
Generates the required folder structure for an R project.
}
\details{
This function creates the necessary directories such as \code{functions}, \code{scripts}, \code{tests}, \code{outputs}, \code{config}, and \code{report}.
}
