% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_renv.R
\name{initialize_renv}
\alias{initialize_renv}
\title{Initialize renv}
\usage{
initialize_renv(project_path)
}
\arguments{
\item{project_path}{The path to the project directory where renv should be initialized.}
}
\value{
No return value, initializes renv.
}
\description{
Sets up the renv package manager inside the project directory and installs essential packages.
}
\details{
This function initializes renv in the specified project folder and installs common R packages.
}
