% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organise.R
\name{organise}
\alias{organise}
\title{Organise a new R project}
\usage{
organise(
  project_name,
  path = ".",
  project_title = NULL,
  project_description = ""
)
}
\arguments{
\item{project_name}{A string specifying the project name.}

\item{path}{The directory where the project should be created. Default is the current working directory.}
}
\value{
No return value, creates files and directories.
}
\description{
Creates a structured R project with predefined folders, example files, and renv setup.
}
\details{
This function orchestrates the creation of an R project by calling various helper functions.
}
\examples{
organise("my_project")
}
